IR is currently commented out, and changed wireless to uint16

+++ grid_wall_sensor/wireless1.h
@@ -6,6 +6,6 @@ extern RF24 radio;
 
 void wireless_setup(uint8_t wr);
 
-bool wireless_send(uint8_t * buf, uint8_t buf_size);
+bool wireless_send(uint16_t * buf, uint16_t buf_size);
 
-bool wireless_read(uint8_t * buf, uint8_t rd_length);
+bool wireless_read(uint16_t * buf, uint16_t rd_length);
